# frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "3X-UI Panel API",
    "version": "3.x",
    "description": "Programmatic interface to a 3X-UI panel. Authenticate either by logging in (cookie) or with an API token from Settings → Security → API Token (Bearer). All endpoints under /panel/api/* honour both modes — an API token is a full-admin credential, so treat it like the panel password."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current panel (basePath aware)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token from Settings → Security → API Token. Send as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "3x-ui",
        "description": "Session cookie set by POST /login. Browser-only."
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/feed/{token}": {
      "get": {
        "tags": [
          "Subconverter"
        ],
        "summary": "Return the managed Mihomo YAML for a public token.",
        "operationId": "get_feed_token",
        "description": "Unauthenticated public feed endpoint. When the immediate client IP is a configured trusted proxy, X-Real-IP, X-Forwarded-For, X-Forwarded-Proto, and X-Forwarded-Host are honoured for client IP and feed URL generation; otherwise forwarded headers are ignored.",
        "security": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "description": "32-character hexadecimal subscription token.",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mihomo YAML subscription.",
            "content": {
              "application/x-yaml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Client IP could not be determined.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Request blocked by user-agent policy or IP limit.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Subscription token is invalid, missing, disabled, or hidden by user-agent policy.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Subscription lookup, IP enforcement, or user-agent policy check failed.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/feed/{token}/nodes": {
      "get": {
        "tags": [
          "Subconverter"
        ],
        "summary": "Return the proxy-provider node list for a public token.",
        "operationId": "get_feed_token_nodes",
        "description": "Unauthenticated public provider endpoint. It checks the same token, user-agent, and trusted-proxy header rules as /feed/:token, but does not bind a new IP slot.",
        "security": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "description": "32-character hexadecimal subscription token.",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mihomo proxy-provider YAML.",
            "content": {
              "application/x-yaml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Client IP could not be determined.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Request blocked by user-agent policy or IP limit.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Subscription token is invalid, missing, disabled, or hidden by user-agent policy.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Subscription lookup, IP check, rendering, or user-agent policy check failed.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}